requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
python-dateutil==2.5.3
redis==2.10.5
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
